# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Desk API",
    "version": "0.0.0",
    "description": "Generated from packages/contracts/src via scripts/openapi.ts. Route list is hand-mapped from specs/001-phased-product-baseline/contracts/api.md; schemas are auto-converted from zod."
  },
  "paths": {
...
    "/widgets": {
      "get": {
        "summary": "List dashboard widgets with figures",
        "responses": {
          "200": {
            "description": "List dashboard widgets with figures",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WidgetsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a widget",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WidgetCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Add a widget",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WidgetResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/widgets/{id}": {
      "patch": {
        "summary": "Change a widget",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WidgetPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Change a widget",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WidgetResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove a widget",
        "responses": {
          "default": {
            "description": "See specs/001-phased-product-baseline/contracts/api.md"
          }
        }
      }
    },
    "/widgets/order": {
      "put": {
        "summary": "Reorder widgets",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reorder widgets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WidgetsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OrderBody": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["ids"],
        "additionalProperties": false
      },
...
      "WidgetCreate": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": ["currency", "weather", "sunrise", "spend_pace", "fixed_costs"]
          },
          "settings": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          },
          "placeId": {
            "type": "string"
          },
          "place": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "minLength": 1,
                "maxLength": 100
              },
              "admin1": {
                "anyOf": [
                  {
                    "type": "string",
                    "maxLength": 100
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "country": {
                "type": "string",
                "minLength": 1,
                "maxLength": 100
              },
              "lat": {
                "type": "number",
                "minimum": -90,
                "maximum": 90
              },
              "lon": {
                "type": "number",
                "minimum": -180,
                "maximum": 180
              },
              "timeZone": {
                "type": "string",
                "maxLength": 64
              }
            },
            "required": ["name", "admin1", "country", "lat", "lon", "timeZone"],
            "additionalProperties": false
          },
          "duplicateOf": {
            "type": "string"
          }
        },
        "required": ["kind"],
        "additionalProperties": false
      },
...
      "WidgetPatch": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "settings": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          },
          "placeId": {
            "type": "string"
          },
          "place": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "minLength": 1,
                "maxLength": 100
              },
              "admin1": {
                "anyOf": [
                  {
                    "type": "string",
                    "maxLength": 100
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "country": {
                "type": "string",
                "minLength": 1,
                "maxLength": 100
              },
              "lat": {
                "type": "number",
                "minimum": -90,
                "maximum": 90
              },
              "lon": {
                "type": "number",
                "minimum": -180,
                "maximum": 180
              },
              "timeZone": {
                "type": "string",
                "maxLength": 64
              }
            },
            "required": ["name", "admin1", "country", "lat", "lon", "timeZone"],
            "additionalProperties": false
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}